backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio[asgi]
boto3
amazon-transcribe
aiohttp
awscrt
python-dotenv